+++ google-cloud-spanner/lib/google/cloud/spanner/session.rb
@@ -1488,17 +1488,24 @@ module Google
         # @param exclude_txn_from_change_streams [::Boolean] Optional. Defaults to `false`.
         #   When `exclude_txn_from_change_streams` is set to `true`, it prevents read
         #   or write transactions from being tracked in change streams.
+        # @param request_options [::Hash, nil] Optional. Common request options.
+        #   Example option: `:priority`.
         # @private
         # @return [::Google::Cloud::Spanner::Transaction]
-        def create_transaction exclude_txn_from_change_streams: false, read_lock_mode: nil
+        def create_transaction exclude_txn_from_change_streams: false, read_lock_mode: nil,
+                               request_options: nil
           route_to_leader = LARHeaders.begin_transaction true
+          request_options = Convert.to_request_options request_options, tag_type: :transaction_tag
+          transaction_tag = request_options[:transaction_tag] if request_options
           tx_grpc = service.begin_transaction path,
                                               route_to_leader: route_to_leader,
                                               exclude_txn_from_change_streams: exclude_txn_from_change_streams,
+                                              request_options: request_options,
                                               read_lock_mode: read_lock_mode
           Transaction.from_grpc \
             tx_grpc, self,
-            exclude_txn_from_change_streams: exclude_txn_from_change_streams, read_lock_mode: read_lock_mode
+            exclude_txn_from_change_streams: exclude_txn_from_change_streams, read_lock_mode: read_lock_mode,
+            transaction_tag: transaction_tag
         end
 
         # Creates a new empty transaction wrapper without a server-side object.
@@ -1512,12 +1519,28 @@ module Google
         #   An id of the previous transaction, if this new transaction wrapper is being created
         #   as a part of a retry. Previous transaction id should be added to TransactionOptions
         #   of a new ReadWrite transaction when retry is attempted.
+        # @param transaction_tag [::String, nil] Optional.
+        #   A tag used for statistics collection about this transaction.
+        # @param read_lock_mode [::Symbol, nil] Optional. The read lock mode for the transaction.
+        #   Can be one of the following:
+        #   * `:READ_LOCK_MODE_UNSPECIFIED` : The default unspecified read lock mode.
+        #   * `:PESSIMISTIC` : The pessimistic lock mode, where depending on the isolation level and/or lock
+        #       requested, locks are acquired on read.
+        #   * `:OPTIMISTIC` : The optimistic lock mode, where locks are not acquired on read. Depending on the
+        #       isolation level and/or lock requested on a read, data may be validated at commit time to be not
+        #       changed since the transaction started.
+        # @param isolation_level [::Symbol, nil] Optional. The isolation level for the transaction.
+        #   Can be one of the following:
+        #   * `:ISOLATION_LEVEL_UNSPECIFIED` : The default unspecified isolation level.
+        #   * `:SERIALIZABLE` : The serializable isolation level.
+        #   * `:REPEATABLE_READ` : The repeatable read isolation level.
         # @private
         # @return [::Google::Cloud::Spanner::Transaction] The new *empty-wrapper* transaction object.
         def create_empty_transaction exclude_txn_from_change_streams: false, previous_transaction_id: nil,
-                                     read_lock_mode: nil
+                                     read_lock_mode: nil, transaction_tag: nil, isolation_level: nil
           Transaction.from_grpc nil, self, exclude_txn_from_change_streams: exclude_txn_from_change_streams,
-previous_transaction_id: previous_transaction_id, read_lock_mode: read_lock_mode
+                                previous_transaction_id: previous_transaction_id, read_lock_mode: read_lock_mode,
+                                transaction_tag: transaction_tag, isolation_level: isolation_level
         end
 
         # If the session is non-multiplexed, keeps the session alive by executing `"SELECT 1"`.